Fix:검색 선택된 키워드 1개일 때 끝에 boolean 뜨는 버그 수정
